Added GenAI warning and info
