+++ src/components/NavItems/tools/C2pa/Hooks/useGetC2paData.jsx
@@ -1,6 +1,7 @@
 import {
   createC2pa,
   selectEditsAndActivity,
+  selectGenerativeInfo,
   selectProducer,
   selectSocialAccounts,
 } from "c2pa";
@@ -89,7 +90,11 @@ async function readManifest(manifest, parent, result, url, depth) {
     };
 
     const editsAndActivity = await selectEditsAndActivity(manifest);
+
     if (editsAndActivity) manifestData.editsAndActivity = editsAndActivity;
+
+    const generativeInfo = await selectGenerativeInfo(manifest);
+    if (generativeInfo) manifestData.generativeInfo = generativeInfo;
 
     const captureInfo = exifData(manifest.assertions.data);
     if (captureInfo) manifestData.captureInfo = captureInfo;
@@ -200,7 +205,7 @@ async function loadTrustResource(file) {
  *
  * @returns {Object} settings allowing the C2PA. Read function to determine if the source of the Content Credentials is on adobe's trusted list
  */
-async function getToolkitSettings() {
+export async function getToolkitSettings() {
   const [trustAnchors, allowedList, trustConfig] = await Promise.all(
     ["anchors.pem", "allowed.sha256.txt", "store.cfg"].map(loadTrustResource),
   );
