+++ src/components/NavItems/tools/SyntheticImageDetection/index.jsx
@@ -1,86 +1,36 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useLocation } from "react-router-dom";
 
 import Alert from "@mui/material/Alert";
-import Box from "@mui/material/Box";
-import Card from "@mui/material/Card";
-import FormControlLabel from "@mui/material/FormControlLabel";
-import FormGroup from "@mui/material/FormGroup";
-import LinearProgress from "@mui/material/LinearProgress";
 import Stack from "@mui/material/Stack";
-import Switch from "@mui/material/Switch";
 
 import { Gradient } from "@mui/icons-material";
 
+import SyntheticImageDetectionForm from "@/components/NavItems/tools/SyntheticImageDetection/SyntheticImageDetectionForm";
 import { useSyntheticImageDetection } from "@/components/NavItems/tools/SyntheticImageDetection/useSyntheticImageDetection";
-import { ROLES } from "@/constants/roles";
-import {
-  resetSyntheticImageDetectionImage,
-  setSyntheticImageDetectionLoading,
-} from "@/redux/actions/tools/syntheticImageDetectionActions";
-import {
-  preprocessFileUpload,
-  resizeImageWithWorker,
-} from "@Shared/Utils/fileUtils";
+import { resetSyntheticImageDetectionImage } from "@/redux/actions/tools/syntheticImageDetectionActions";
 import { i18nLoadNamespace } from "components/Shared/Languages/i18nLoadNamespace";
-import { setError } from "redux/reducers/errorReducer";
 
 import HeaderTool from "../../../Shared/HeaderTool/HeaderTool";
-import StringFileUploadField from "../../../Shared/StringFileUploadField";
 import SyntheticImageDetectionResults from "./syntheticImageDetectionResults";
 
 const SyntheticImageDetection = () => {
-  const location = useLocation();
-  const urlParams = new URLSearchParams(location.search);
-  const urlParam = urlParams.get("url");
-
-  const keyword = i18nLoadNamespace(
-    "components/NavItems/tools/SyntheticImageDetection",
-  );
   const keywordAllTools = i18nLoadNamespace(
     "components/NavItems/tools/Alltools",
   );
+
   const keywordWarning = i18nLoadNamespace("components/Shared/OnWarningInfo");
 
-  const role = useSelector((state) => state.userSession.user.roles);
-
-  const isLoading = useSelector(
-    (state) => state.syntheticImageDetection.loading,
-  );
   const result = useSelector((state) => state.syntheticImageDetection.result);
-  const url = useSelector((state) => state.syntheticImageDetection.url);
   const nd = useSelector((state) => state.syntheticImageDetection.duplicates);
-  const [input, setInput] = useState(url ? url : "");
-  const [imageFile, setImageFile] = useState(undefined);
-
-  const [imageType, setImageType] = useState(undefined);
-
-  const [autoResizeLocalFile, setAutoResizeLocalFile] = useState(true);
+  const url = useSelector((state) => state.syntheticImageDetection.url);
 
   const dispatch = useDispatch();
 
-  const IMAGE_FROM = {
-    URL: "url",
-    UPLOAD: "local",
-  };
+  const [imageType, setImageType] = useState(undefined);
 
-  const { startDetection } = useSyntheticImageDetection({
-    dispatch,
-  });
-
-  useEffect(() => {
-    if (urlParam && !input) {
-      setInput(urlParam);
-      handleSubmit(urlParam);
-    }
-  }, []);
-
-  useEffect(() => {
-    if (url && input && !result) {
-      handleSubmit(input);
-    }
-  }, [url, input, result]);
+  const [input, setInput] = useState(url ? url : "");
+  const [imageFile, setImageFile] = useState(undefined);
 
   const resetState = () => {
     setInput("");
@@ -88,59 +38,9 @@ const SyntheticImageDetection = () => {
     dispatch(resetSyntheticImageDetectionImage());
   };
 
-  const preprocessingSuccess = (file) => {
-    setImageFile(file);
-    return file;
-  };
-
-  const preprocessingError = () => {
-    dispatch(setError(keywordWarning("warning_file_too_big")));
-  };
-
-  const preprocessImage = (file) => {
-    return preprocessFileUpload(
-      file,
-      role,
-      undefined,
-      preprocessingSuccess,
-      preprocessingError,
-    );
-  };
-
-  /**
-   *
-   * @param url {string}
-   * @returns {Promise<void>}
-   */
-  const handleSubmit = async (url) => {
-    const processedFile =
-      autoResizeLocalFile && imageFile
-        ? role.includes(ROLES.BETA_TESTER)
-          ? await resizeImageWithWorker(imageFile)
-          : imageFile
-        : imageFile;
-
-    if (autoResizeLocalFile && processedFile) {
-      setImageFile(processedFile);
-    }
-
-    dispatch(resetSyntheticImageDetectionImage());
-
-    const urlInput = url ? url : input;
-
-    const type =
-      urlInput && typeof urlInput === "string"
-        ? IMAGE_FROM.URL
-        : IMAGE_FROM.UPLOAD;
-
-    dispatch(setSyntheticImageDetectionLoading(true));
-
-    await startDetection({
-      type: type,
-      url: urlInput,
-      imageFile: processedFile,
-    });
-  };
+  const { startDetection, c2paData } = useSyntheticImageDetection({
+    dispatch,
+  });
 
   useEffect(() => {
     if (!result) return;
@@ -173,12 +73,8 @@ const SyntheticImageDetection = () => {
     }
   }, [imageFile, input, result]);
 
-  const toggleAutoResizeLocalFile = () => {
-    setAutoResizeLocalFile((prev) => !prev);
-  };
-
   return (
-    <Box>
+    <Stack direction="column" spacing={4}>
       <HeaderTool
         name={keywordAllTools("navbar_synthetic_image_detection")}
         description={keywordAllTools(
@@ -194,78 +90,20 @@ const SyntheticImageDetection = () => {
           />
         }
       />
-      <Stack
-        direction="column"
-        spacing={2}
-        sx={{
-          mb: 4,
-        }}
-      >
-        <Alert severity="warning">
-          {keywordWarning("warning_beta_synthetic_image_detection")}
-        </Alert>
-      </Stack>
-      <Card variant="outlined">
-        <Box
-          sx={{
-            p: 4,
-          }}
-        >
-          <Stack direction="column" spacing={2}>
-            <form>
-              <StringFileUploadField
-                labelKeyword={keyword("synthetic_image_detection_link")}
-                placeholderKeyword={keyword(
-                  "synthetic_image_detection_placeholder",
-                )}
-                submitButtonKeyword={keyword("submit_button")}
-                localFileKeyword={keyword("button_localfile")}
-                urlInput={input}
-                setUrlInput={setInput}
-                fileInput={imageFile}
-                setFileInput={setImageFile}
-                handleSubmit={handleSubmit}
-                fileInputTypesAccepted={"image/*"}
-                handleCloseSelectedFile={resetState}
-                preprocessLocalFile={preprocessImage}
-                isParentLoading={isLoading}
-                handleClearUrl={resetState}
-              />
-            </form>
 
-            {role.includes(ROLES.BETA_TESTER) && imageFile && (
-              <FormGroup>
-                <FormControlLabel
-                  control={
-                    <Switch
-                      checked={autoResizeLocalFile}
-                      onChange={toggleAutoResizeLocalFile}
-                      size="small"
-                      disabled={isLoading}
-                    />
-                  }
-                  label="Auto-Resize"
-                />
-              </FormGroup>
-            )}
+      <Alert severity="warning">
+        {keywordWarning("warning_beta_synthetic_image_detection")}
+      </Alert>
 
-            {isLoading && (
-              <Box
-                sx={{
-                  mt: 3,
-                }}
-              >
-                <LinearProgress />
-              </Box>
-            )}
-          </Stack>
-        </Box>
-      </Card>
-      <Box
-        sx={{
-          m: 3,
-        }}
+      <SyntheticImageDetectionForm
+        resetState={resetState}
+        input={input}
+        setInput={setInput}
+        imageFile={imageFile}
+        setImageFile={setImageFile}
+        startDetection={startDetection}
       />
+
       {result && (
         <SyntheticImageDetectionResults
           results={result}
@@ -273,9 +111,10 @@ const SyntheticImageDetection = () => {
           handleClose={resetState}
           nd={nd}
           imageType={imageType}
+          c2paData={c2paData}
         />
       )}
-    </Box>
+    </Stack>
   );
 };
